+++ python/cugraph/link_analysis/pagerank_wrapper.pyx
@@ -44,7 +44,7 @@ def pagerank(input_graph, alpha=0.85, personalization=None, max_iter=100, tol=1.
     [weights] = graph_wrapper.datatype_cast([input_graph.transposedadjlist.weights], [np.float32, np.float64])
 
     num_verts = input_graph.number_of_vertices()
-    num_edges = input_graph.number_of_edges()
+    num_edges = len(indices)
 
     df = cudf.DataFrame()
     df['vertex'] = cudf.Series(np.zeros(num_verts, dtype=np.int32))