+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # cuGraph 0.13.0 (Date TBD)
 
 ## New Features
+- PR #736 cuHornet KTruss integration
 - PR #735 Integration gunrock's betweenness centrality
 
 ## Improvements
@@ -17,6 +18,8 @@
 - PR #738 Move tests directory up a level
 - PR #739 Updated Notebooks
 - PR #740 added utility to extract paths from SSSP/BFS results
+- PR #742 Rremove gdf column from jaccard
+- PR #741 Added documentation for running and adding new benchmarks and shell script to automate
 - PR #747 updated viewing of graph, datatypecasting and two hop neighbor unrenumbering for multi column
 
 ## Bug Fixes
@@ -27,6 +30,8 @@
 - PR #726 Fixed SSSP notebook issues in last cell
 - PR #728 Temporary fix for dask attribute error issue
 - PR #733 Fixed multi-column renumbering issues with indexes
+- PR #746 Dask + Distributed 2.12.0+
+- PR #753 ECG Error
 
 # cuGraph 0.12.0 (04 Feb 2020)